Doc doesn't mention "complete migration"
If the VM is being migrated to Azure, Complete Migration should be used after the validation is successful. If the intent is to bring it back on-premises (failover from Azure to on-premises) at a later point, then Commit should be used.

added this in where appropriate.  May be required in other docs (Hyper-V/VMM)